+++ src/app/search/query-parser.test.ts
@@ -50,6 +50,11 @@ const cases = [
   [
     '  (\n    (\n    (is:weapon -is:maxpower powerlimit:1060 or tag:junk or is:blue)\n    or\n    (\n    (is:armor -is:exotic -is:classitem)\n  \n    -(is:titan (basestat:recovery:>=18 or basestat:total:>=63))\n    -(is:hunter ((basestat:recovery:>=13 basestat:mobility:>=18) or basestat:recovery:>15 or basestat:total:>=63))\n    -(is:warlock ((basestat:recovery:>=18 discipline:>=17) or basestat:total:>=63))\n  \n    -(\n    ((basestat:mobility:>=18 basestat:resilience:>=13) or\n    (basestat:mobility:>=18 basestat:recovery:>=13) or\n    (basestat:mobility:>=18 basestat:discipline:>=13) or\n    (basestat:mobility:>=18 basestat:intellect:>=13) or\n    (basestat:mobility:>=18 basestat:strength:>=13) or\n    (basestat:resilience:>=18 basestat:mobility:>=13) or\n    (basestat:resilience:>=18 basestat:recovery:>=13) or\n    (basestat:resilience:>=18 basestat:discipline:>=13) or\n    (basestat:resilience:>=18 basestat:intellect:>=13) or\n    (basestat:resilience:>=18 basestat:strength:>=13) or\n    (basestat:recovery:>=18 basestat:mobility:>=13) or\n    (basestat:recovery:>=18 basestat:resilience:>=13) or\n    (basestat:recovery:>=18 basestat:discipline:>=13) or\n    (basestat:recovery:>=18 basestat:intellect:>=13) or\n    (basestat:recovery:>=18 basestat:strength:>=13) or\n    (basestat:discipline:>=18 basestat:mobility:>=13) or\n    (basestat:discipline:>=18 basestat:resilience:>=13) or\n    (basestat:discipline:>=18 basestat:recovery:>=13) or\n    (basestat:discipline:>=18 basestat:intellect:>=13) or\n    (basestat:discipline:>=18 basestat:strength:>=13) or\n    (basestat:intellect:>=18 basestat:mobility:>=13) or\n    (basestat:intellect:>=18 basestat:resilience:>=13) or\n    (basestat:intellect:>=18 basestat:recovery:>=13) or\n    (basestat:intellect:>=18 basestat:discipline:>=13) or\n    (basestat:intellect:>=18 basestat:strength:>=13) or\n    (basestat:strength:>=18 basestat:mobility:>=13) or\n    (basestat:strength:>=18 basestat:resilience:>=13) or\n    (basestat:strength:>=18 basestat:recovery:>=13) or\n    (basestat:strength:>=18 basestat:discipline:>=13) or\n    (basestat:strength:>=18 basestat:intellect:>=13))\n    )\n  \n    -(\n    ((basestat:mobility:>=13 basestat:resilience:>=13 basestat:recovery:>=13) or\n    (basestat:mobility:>=13 basestat:resilience:>=13 basestat:discipline:>=13) or\n    (basestat:mobility:>=13 basestat:resilience:>=13 basestat:intellect:>=13) or\n    (basestat:mobility:>=13 basestat:resilience:>=13 basestat:strength:>=13) or\n    (basestat:mobility:>=13 basestat:recovery:>=13 basestat:discipline:>=13) or\n    (basestat:mobility:>=13 basestat:recovery:>=13 basestat:intellect:>=13) or\n    (basestat:mobility:>=13 basestat:recovery:>=13 basestat:strength:>=13) or\n    (basestat:mobility:>=13 basestat:discipline:>=13 basestat:intellect:>=13) or\n    (basestat:mobility:>=13 basestat:discipline:>=13 basestat:strength:>=13) or\n    (basestat:mobility:>=13 basestat:intellect:>=13 basestat:strength:>=13) or\n    (basestat:resilience:>=13 basestat:recovery:>=13 basestat:discipline:>=13) or\n    (basestat:resilience:>=13 basestat:recovery:>=13 basestat:intellect:>=13) or\n    (basestat:resilience:>=13 basestat:recovery:>=13 basestat:strength:>=13) or\n    (basestat:resilience:>=13 basestat:discipline:>=13 basestat:intellect:>=13) or\n    (basestat:resilience:>=13 basestat:discipline:>=13 basestat:strength:>=13) or\n    (basestat:resilience:>=13 basestat:intellect:>=13 basestat:strength:>=13) or\n    (basestat:recovery:>=13 basestat:discipline:>=13 basestat:intellect:>=13) or\n    (basestat:recovery:>=13 basestat:discipline:>=13 basestat:strength:>=13) or\n    (basestat:recovery:>=13 basestat:intellect:>=13 basestat:strength:>=13) or\n    (basestat:discipline:>=13 basestat:intellect:>=13 basestat:strength:>=13))\n    )\n  \n    -(basestat:mobility:>=8 basestat:resilience:>=8 basestat:recovery:>=8 basestat:discipline:>=8 basestat:intellect:>=8 basestat:strength:>=8)\n    )\n  \n    or\n    (is:classitem ((is:dupelower -is:modded) or (is:sunset))) \n    or\n    (is:armor -powerlimit:>1060) \n    or\n    (is:armor is:blue)\n    )\n    -tag:keep -tag:archive -tag:favorite -tag:infuse -is:maxpower -power:>=1260 -is:inloadout -is:masterwork\n    )',
   ],
+  // Plaintext special case
+  ['not forgotten'],
+  ['not (forgotten)'],
+  ['not "forgotten"'],
+  ['gnawing hunger'],
 ];
 
 // Each of these asserts that the first query is the same as the second query once parsed
@@ -58,8 +63,8 @@ const equivalentSearches = [
     'is:blue is:weapon or is:armor not:maxpower',
     'is:blue and (is:weapon or is:armor) and -is:maxpower',
   ],
-  ['not forgotten', "-'forgotten'"],
-  ['cluster tracking', '"cluster" and "tracking"'],
+  ['not forgotten', '"not forgotten"'],
+  ['cluster tracking', '"cluster tracking"'],
   [
     'is:weapon and is:sniperrifle or not is:armor and modslot:arrival',
     '(is:weapon and is:sniperrifle) or (-is:armor and modslot:arrival)',
@@ -92,7 +97,7 @@ const canonicalize = [
     'is:rocketlauncher (perk:"tracking module" or perk:cluster)',
   ],
   ['( power:>1000 and -modslot:arrival ) ', '-modslot:arrival power:>1000'],
-  ['food fight', 'fight food'],
+  ['food fight', '"food fight"'],
 ];
 
 test.each(cases)('parse |%s|', (query) => {