+++ src/app/search/query-parser.ts
@@ -273,6 +273,11 @@ const bareWords = /[^\s)]+/y;
 // Whitespace that doesn't match anything else is an implicit `and`
 const whitespace = /\s+/y;
 
+// A search matching this regex is treated as a single quoted string, to handle the common case of
+// someone typing "gnawing hunger" or "kill clip". This also means "not forgotten" is different from
+// "not (forgotten)" or "not "forgotten"".
+const isPlainSearch = /^[^":()<=>-]+$/;
+
 /**
  * The lexer yields a series of tokens representing the linear structure of the search query.
  * This throws an exception if it finds an invalid input.
@@ -286,6 +291,11 @@ export function* lexer(query: string): Generator<Token> {
   // http://blog.tatedavies.com/2012/08/28/replace-microsoft-chars-in-javascript/
   query = query.replace(/[\u2018|\u2019|\u201A]/g, "'");
   query = query.replace(/[\u201C|\u201D|\u201E]/g, '"');
+
+  // First test for the simple case of plaintext search
+  if (isPlainSearch.test(query)) {
+    return yield ['filter', 'keyword', query];
+  }
 
   let match: string | undefined;
   let i = 0;